more classes moved to platform-gui
