+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/hashgraph/internal/HashgraphGuiControls.java
@@ -17,9 +17,10 @@
 package com.swirlds.platform.gui.hashgraph.internal;
 
 import static com.swirlds.gui.GuiUtils.wrap;
-import static com.swirlds.platform.gui.hashgraph.internal.HashgraphGuiUtils.DEFAULT_GENERATIONS_TO_DISPLAY;
+import static com.swirlds.gui.hashgraph.HashgraphGuiConstants.DEFAULT_GENERATIONS_TO_DISPLAY;
 
 import com.swirlds.gui.GuiUtils;
+import com.swirlds.gui.hashgraph.HashgraphPictureOptions;
 import com.swirlds.platform.consensus.GraphGenerations;
 import com.swirlds.platform.event.EventConstants;
 import java.awt.Checkbox;
