+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/hashgraph/internal/ShadowgraphGuiSource.java
@@ -17,11 +17,11 @@
 package com.swirlds.platform.gui.hashgraph.internal;
 
 import com.swirlds.common.system.events.PlatformEvent;
+import com.swirlds.gui.hashgraph.HashgraphGuiSource;
 import com.swirlds.platform.consensus.GraphGenerations;
 import com.swirlds.platform.event.EventUtils;
 import com.swirlds.platform.gossip.shadowgraph.ShadowEvent;
 import com.swirlds.platform.gossip.shadowgraph.ShadowGraph;
-import com.swirlds.platform.gui.hashgraph.HashgraphGuiSource;
 import com.swirlds.platform.internal.EventImpl;
 
 /**
