+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/internal/WinTab2Consensus.java
@@ -18,6 +18,7 @@ package com.swirlds.platform.gui.internal;
 
 import static com.swirlds.gui.GuiUtils.wrap;
 
+import com.swirlds.gui.GuiUtils;
 import com.swirlds.gui.PrePaintableJPanel;
 import com.swirlds.platform.Consensus;
 import com.swirlds.platform.SwirldsPlatform;
@@ -43,7 +44,7 @@ class WinTab2Consensus extends PrePaintableJPanel {
      * Instantiate and initialize content of this tab.
      */
     public WinTab2Consensus() {
-        text = WinBrowser.newJTextArea();
+        text = GuiUtils.newJTextArea("");
         add(text);
     }
