+++ platform-sdk/swirlds-platform-gui/src/main/java/module-info.java
@@ -1,6 +1,8 @@
 module com.swirlds.gui {
     exports com.swirlds.gui;
+    exports com.swirlds.gui.hashgraph;
 
     requires transitive java.desktop;
     requires org.apache.commons.lang3;
+    requires com.swirlds.common;
 }
