+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/gui/HashgraphGuiTest.java
@@ -16,7 +16,7 @@
 
 package com.swirlds.platform.test.gui;
 
-import com.swirlds.platform.gui.hashgraph.HashgraphGuiSource;
+import com.swirlds.gui.hashgraph.HashgraphGuiSource;
 import com.swirlds.platform.gui.hashgraph.internal.FinalShadowgraphGuiSource;
 import com.swirlds.platform.test.consensus.TestIntake;
 import com.swirlds.platform.test.event.generator.StandardGraphGenerator;
